requirements.txt
cloudpickle==3.1.2
contourpy==1.3.3
cycler==0.12.1
dask==2025.11.0
dm-haiku==0.0.15
fonttools==4.60.1
fsspec==2025.10.0
jax==0.8.0
jax-cuda12-pjrt==0.8.0
jax-cuda12-plugin==0.8.0
jaxlib==0.8.0
jmp==0.0.4
jraph==0.0.6.dev0
kiwisolver==1.4.9
locket==1.0.0
matplotlib==3.10.7
ml_dtypes==0.5.3
netCDF4==1.7.3
numpy==2.3.4
nvidia-cublas-cu12==12.9.1.4
nvidia-cuda-cupti-cu12==12.9.79
nvidia-cuda-nvcc-cu12==12.9.86
nvidia-cuda-nvrtc-cu12==12.9.86
nvidia-cuda-runtime-cu12==12.9.79
nvidia-cudnn-cu12==9.15.0.57
nvidia-cufft-cu12==11.4.1.4
nvidia-cusolver-cu12==11.7.5.82
nvidia-cusparse-cu12==12.5.10.65
nvidia-nccl-cu12==2.28.7
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvshmem-cu12==3.4.5
opt_einsum==3.4.0
optax==0.2.6
packaging==25.0
pandas==2.3.3
partd==1.4.2
pillow==12.0.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
scipy==1.16.3
setuptools==80.9.0
six==1.17.0
tabulate==0.9.0
toolz==1.1.0
typing_extensions==4.15.0
tzdata==2025.2
uv==0.9.7
wheel==0.45.1
xarray==2025.10.1